correcoes  + refatoracao de interfaces
