+++ src/components/screens/PreRecebimento/DocumentosRecebimento/index.tsx
@@ -6,11 +6,9 @@ import { listarDocumentosRecebimento } from "../../../../services/documentosRece
 import Filtros from "./components/Filtros";
 import Listagem from "./components/Listagem";
 import { Paginacao } from "../../../Shareable/Paginacao";
-import {
-  DocumentosRecebimento,
-  FiltrosDocumentosRecebimento,
-} from "./interfaces";
+import { FiltrosDocumentosRecebimento } from "./interfaces";
 import { ResponseDocumentosRecebimento } from "interfaces/responses.interface";
+import { DocumentosRecebimento } from "interfaces/pre_recebimento.interface";
 
 export default () => {
   const [carregando, setCarregando] = useState<boolean>(false);
