+++ src/components/screens/PreRecebimento/DocumentosRecebimento/interfaces.ts
@@ -1,11 +1,3 @@
-export interface CronogramaSimples {
-  uuid: string;
-  nome_produto: string;
-  numero: string;
-  pregao_chamada_publica: string;
-  value?: string;
-}
-
 export interface Arquivo {
   nome: string;
   base64: string;
@@ -50,13 +42,4 @@ export interface FiltrosDocumentosRecebimento {
   numero_cronograma?: string;
   status?: string;
   data_criacao?: string;
-}
-
-export interface DocumentosRecebimento {
-  criado_em: string;
-  nome_produto: string;
-  numero_cronograma: string;
-  pregao_chamada_publica: string;
-  status: string;
-  uuid: string;
 }
