+++ src/interfaces/produto.interface.ts
@@ -1,0 +1,6 @@
+export interface ProdutoLogistica {
+  criado_em: string;
+  nome: string;
+  uuid: string;
+  status: string;
+}
